# alem-ekub-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habte",
            "email": "habtadugna@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "getPageOfUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseUserResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "operationId": "updateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "request": {
                                        "$ref": "#/components/schemas/UserRequest"
                                    },
                                    "profilePic": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "idCardImg": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "idCardImg",
                                    "profilePic",
                                    "request"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "put": {
                "tags": [
                    "users"
                ],
                "operationId": "updateProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/password/{user-id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "operationId": "updatePassword",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/cancel-invitation/{user-id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "operationId": "cancelInvitation",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ekub-id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/users/profile-picture": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "uploadProfilePicture",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/invite/{user-id}": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "inviteUserToEkub",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ekub-id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/id-card-image": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "uploadIdCardImage",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/users/{user-id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/search-to-invite/{ekub-id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "searchUsersToInvite",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search-term",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/invited-users/{ekub-id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "getInvitedUsersInEkubAndNotJoined",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/payments/user/{user-id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "operationId": "getUserPayments",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "ekubIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "remark": {
                        "type": "string"
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "phoneNumber",
                    "profession",
                    "username"
                ]
            },
            "IdResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    }
                }
            },
            "PasswordUpdateRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "oldPassword": {
                        "type": "string"
                    },
                    "newPassword": {
                        "type": "string"
                    }
                },
                "required": [
                    "newPassword",
                    "oldPassword",
                    "username"
                ]
            },
...
            "PageResponseUserResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "profilePicUrl": {
                        "type": "string"
                    },
                    "idCardImageUrl": {
                        "type": "string"
                    },
                    "guarantors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "ekubs": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ekubIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
...
            "RoundResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "ekubName": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roundNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "winner": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "paid": {
                        "type": "boolean"
                    },
                    "userGuarantees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserGuaranteeResponse"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UserGuaranteeResponse": {
                "type": "object",
                "properties": {
                    "guarantorId": {
                        "type": "string"
                    },
                    "guarantor": {
                        "type": "string"
                    },
                    "guaranteedId": {
                        "type": "string"
                    },
                    "guaranteed": {
                        "type": "string"
                    }
                }
            },
...
            "PaymentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "ekubName": {
                        "type": "string"
                    },
                    "round": {
                        "$ref": "#/components/schemas/RoundResponse"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}